+++ contracts/vm_api/transaction.cpp
@@ -6,7 +6,7 @@
 
 #include "vm_api.h"
 
-void send_deferred(const uint128_t& sender_id, uint64_t payer, const char *serialized_transaction, size_t size, uint32_t replace_existing) {
+void send_deferred(const uint128_t& sender_id, uint64_t payer, const char *serialized_transaction, uint32_t size, uint32_t replace_existing) {
    get_vm_api()->send_deferred(&sender_id, payer,serialized_transaction, size, replace_existing);
 }
 
@@ -14,7 +14,7 @@ int cancel_deferred(const uint128_t& sender_id) {
    return get_vm_api()->cancel_deferred(&sender_id);
 }
 
-void _send_deferred(const uint128_t* sender_id, uint64_t payer, const char *serialized_transaction, size_t size, uint32_t replace_existing) {
+void _send_deferred(const uint128_t* sender_id, uint64_t payer, const char *serialized_transaction, uint32_t size, uint32_t replace_existing) {
    get_vm_api()->send_deferred(sender_id, payer,serialized_transaction, size, replace_existing);
 }
 
@@ -22,11 +22,11 @@ int _cancel_deferred(const uint128_t* sender_id) {
    return get_vm_api()->cancel_deferred(sender_id);
 }
 
-size_t read_transaction(char *buffer, size_t size) {
+uint32_t read_transaction(char *buffer, uint32_t size) {
    return get_vm_api()->read_transaction(buffer, size);
 }
 
-size_t transaction_size() {
+uint32_t transaction_size() {
    return get_vm_api()->transaction_size();
 }
 
@@ -42,7 +42,7 @@ _time expiration() {
    return get_vm_api()->expiration();
 }
 
-int get_action( uint32_t type, uint32_t index, char* buff, size_t size ) {
+int get_action( uint32_t type, uint32_t index, char* buff, uint32_t size ) {
    return get_vm_api()->get_action( type, index, buff, size );
 }
 
@@ -54,6 +54,6 @@ void assert_context_free() {
    get_vm_api()->assert_context_free();
 }
 
-int get_context_free_data( uint32_t index, char* buff, size_t size ) {
+int get_context_free_data( uint32_t index, char* buff, uint32_t size ) {
    return get_vm_api()->get_context_free_data( index, buff, size );
 }
